------------------------------------------------------------------------ r3380 | coomi01 | 2010-08-03 04:00:16 -0400 (Tue, 03 Aug 2010) | 23 lines
Cross integration of change 506329

Review Board IP  http://reviewboard.ingres.prv/r/5046

Issue 145993

Bug 124132 - select ... order by next value for <sequence> causes a
segmentation fault.

'next value' is not allowed in an order by clause, and when this is
detected the parser should fail with an error.  Due to the tweaking
of the pss_qualdepth field, continuing can cause a memory overrun
since the save_seq_ops array is only size 100.

The order by processing currently temporarily does this:
cb->pss_qualdepth += (MAXI1 + 1);

which exceeds 100.  This was taken out in main, but it doesn't
change the fact that we don't need to go anywhere near this code
in the first place.

------------------------------------------------------------------------
r3381 | schka01 | 2010-08-03 09:45:14 -0400 (Tue, 03 Aug 2010) | 12 lines

The config.dat upgrade taking dbms.*.direct_io to config.direct_io
was not adding the related direct_io_log and direct_io_load parameters.
If the old dbms.*.direct_io parameter was ON, it makes sense to
have the new direct_io_log ON as well.  (direct_io_load is always
initialized to OFF, as it's much more specialized.)

In addition, have server startup note the direct IO and direct IO
load hint settings into the DBMS log.  (Direct IO log is read
elsewhere and it's more of a nuisance to get at, so I didn't bother.)

------------------------------------------------------------------------