------------------------------------------------------------------------ r3381 | schka01 | 2010-08-03 09:45:14 -0400 (Tue, 03 Aug 2010) | 12 lines
The config.dat upgrade taking dbms.*.direct_io to config.direct_io
was not adding the related direct_io_log and direct_io_load parameters.
If the old dbms.*.direct_io parameter was ON, it makes sense to
have the new direct_io_log ON as well.  (direct_io_load is always
initialized to OFF, as it's much more specialized.)

In addition, have server startup note the direct IO and direct IO
load hint settings into the DBMS log.  (Direct IO log is read
elsewhere and it's more of a nuisance to get at, so I didn't bother.)

------------------------------------------------------------------------
r3382 | hanal04 | 2010-08-03 11:30:09 -0400 (Tue, 03 Aug 2010) | 18 lines

Cross integration of change 506488

Review Board IP  http://reviewboard.ingres.prv/r/5160

The fix for Bug 123136 assumed that a reltid of -1 would
never happen and used this to determine whether writeview()
had been called before. When a an MS-SQL view is registered
in a STAR database and we unload (or copy) the Star Database
writeview() may be called for the first time with a reltid
of -1. As a result we were failing to ensure the text stream
had been opened.
This cahnge adds an explicit "FirstTime" static boolean to
ensure the first call to writeview() will always open the
text stream if required. Further calls will still check
for an open stream if the reltid has changed.

------------------------------------------------------------------------