------------------------------------------------------------------------ r3387 | kiria01 | 2010-08-04 04:45:37 -0400 (Wed, 04 Aug 2010) | 16 lines
Fix SEGV in psl_sscan during normalisation of Unicode.

Background Information:

Reported by customer on 9.3.1
Recreated here on main head revs

Found the ULM corruption and fixed the bug in the 'don't need to normalise'
optimisation. The ULM bug happens when this routine checks for allocated
block size of PSS_SBSIZE but the allocation routine only allocates the
request size (was reserve). The optimisation of returning the raw data
when no net normalisation happened broke when the save_symnext became
stale when a new block was allocated.

------------------------------------------------------------------------