------------------------------------------------------------------------ r3390 | loera01 | 2010-08-04 12:00:11 -0400 (Wed, 04 Aug 2010) | 5 lines
Modified conversion of DAYOFYEAR() scalar to include Gregorian adjustments
for leap year.  Revised CountLiteralItems() to account for scalars that
referenced the target column mulitple times.

------------------------------------------------------------------------
r3391 | horda03 | 2010-08-04 12:15:10 -0400 (Wed, 04 Aug 2010) | 32 lines

Review Board IP  http://reviewboard.ingres.prv/r/5249

Cross integration of change 506556

In the function ValueIsInteger(), when a '-'
is detected, the string is advanced 1 character,
however the length of the string had already
been determined, so the loop that checks that each
character of the (remaining) string is a numeric
now checks the '0' (EOS) character, and hence
returns that the negative number is not a
numeric.

When '-' is detected, decrement the length of the string
to be checked.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------