------------------------------------------------------------------------ r3398 | schka01 | 2010-08-04 22:45:11 -0400 (Wed, 04 Aug 2010) | 8 lines
COPY FROM PROGRAM was silently ignoring the WITH LOG=
error logging instruction.  In an early stage of the 10.0
COPY front-end changes, it seemed like error logging would
be impossible with copying from PROGRAM.  That architecture
decision changed, but the code to turn off error logging
was left in place.  Remove it.

------------------------------------------------------------------------
r3399 | schka01 | 2010-08-05 06:30:31 -0400 (Thu, 05 Aug 2010) | 30 lines

When constant folding was introduced, psl_mk_const_similar
was invented to coerce constants or things that look like
constants (such as parameter markers) to a desired type.
The general notion was that the coercion would be
constant-folded away, resulting in a constant of the
exact desired type.  However, in some circumstances,
the coercion can't be folded away.  A prepared query
being compiled as a REPEATED query via the cache_dynamic
setting is one such circumstance.

psl_mk_const_similar is called with no notion of join ID
context, and so it was creating a coercion operator with
no join ID.  When the parameter marker was actually in
an outer join ON clause, the coercion is created with the
wrong (zero) join ID, and the optimizer complains.

The fix was to allow pst_node to set the proper
join ID for any new operator node depending on the
parser context.  This in turn required that the
PSS_YYVARS context be accessible to pst_node, which
was done by storing a PSS_YYVARS pointer in the
PSS_SES_BLK.  As a byproduct, considerable code in
the parser which computes the join ID can now simply
ask pst_node for the "standard join ID".  Another
byproduct is that the parser action routines
generated by byacc no longer need to pass the PSS_YYVARS
as a routine parameter, they can simply fetch the
pointer from the sess_cb.

------------------------------------------------------------------------