------------------------------------------------------------------------ r3401 | miket | 2010-08-05 13:21:05 -0400 (Thu, 05 Aug 2010) | 16 lines
Trac Number: #584

Bug Number: 122403

Description of changes:
Encryption fixes having mostly to do with feedback from DMT_OPEN so
that the caller can detect an open of a locked encrypted table. Use
this for bailing out of copy (SD 145623) and so that we provide an
appropriate message for an attempt to select from an empty encrypted
table (SD 145619). Collateral to these changes switch terminology
(per QA request) for passphrase activation of encrypted tables from
enabled/disabled (which would seem to apply better to the CREATE TABLE
encryption specifications) to unlocked/locked (i.e. made accessible with
MODIFY ... ENCRYPT WITH PASSPHRASE='foobarkey'). Also cleaned up some
other function documentation and comments.

------------------------------------------------------------------------
r3402 | drivi01 | 2010-08-05 14:15:10 -0400 (Thu, 05 Aug 2010) | 18 lines

There appears to be a refresh problem with the logic
in the DBA Tools installer.  Refresh of the dialog
forces the re-evaluation of all conditions on the dialog
causing the virtual node name to be overwritten with the
hostname.

To make the dialog more user friendly, dialog contains
the logic so that when a user enters hostname and then
clicks tab, the virtual node name is automatically set
to hostname.  However, this is causing an issue when
the dialog is forced to repaint b/c of being forced
to the background and then a forceground.

This change updates the logic in the installer on
hostname field to only set virtual node name to
the hostname if virtual node name is empty.

------------------------------------------------------------------------