------------------------------------------------------------------------ r3402 | drivi01 | 2010-08-05 14:15:10 -0400 (Thu, 05 Aug 2010) | 18 lines
There appears to be a refresh problem with the logic
in the DBA Tools installer.  Refresh of the dialog
forces the re-evaluation of all conditions on the dialog
causing the virtual node name to be overwritten with the
hostname.

To make the dialog more user friendly, dialog contains
the logic so that when a user enters hostname and then
clicks tab, the virtual node name is automatically set
to hostname.  However, this is causing an issue when
the dialog is forced to repaint b/c of being forced
to the background and then a forceground.

This change updates the logic in the installer on
hostname field to only set virtual node name to
the hostname if virtual node name is empty.

------------------------------------------------------------------------
r3403 | drivi01 | 2010-08-05 14:30:13 -0400 (Thu, 05 Aug 2010) | 9 lines

On Upgrade from 9.3.1 to Ingres 10.0, user ends up with 2 sets of utilities
1 is stored under "Utilities" and 1 under "Other Utilities".

This change updates the installer files by adding a "Utilities" directory
to the Directories table and then removing this "Utilities" directory on
install/uninstall, which means "Utilities" will be replaced with
"Other Utilities".

------------------------------------------------------------------------