------------------------------------------------------------------------ r3419 | kiria01 | 2010-08-09 15:45:14 -0400 (Mon, 09 Aug 2010) | 78 lines
Distant join span, nested subqueries suffer from join ID problems

Nested subqueries where the inner qualification is not satisfied
by the immediate containing query gets flattened incorrectly.
This can produce various symptoms such as E_OP0082, E_OP039F and
E_OP0395

Given such a query like:
SELECT * FROM f1, c1 WHERE f1.i = c1.i AND f1.i =
        (SELECT i FROM f2 WHERE f2.i = c1.i AND f2.i =
                (SELECT i FROM f3 WHERE f3.i = c1.i)
        );
the flattening of the innermost subselect would give:
SELECT * FROM f1, c1 WHERE f1.i = c1.i AND f1.i =
        (SELECT f2.i
         FROM   f2 INNER JOIN
                (SELECT f3.i, SINGLETON(f3.i) AS col2
                 FROM f3
                 GROUP BY f3.i) AS dt1
                ON f2.i = SINGLECHK(dt1.col2)
         WHERE f2.i = c1.i AND dt1.i = c1.i
        );

The subsequent flattening should transform into the following
but the code was mishandling the local reference to the re-exported
dt1.i as col2.

SELECT f1.*, c1.*
FROM    f1,
        c1 INNER JOIN
        (SELECT f2.i, dt1.i AS col2, SINGLETON(f2.i) AS col3
         FROM   f2 INNER JOIN
                (SELECT f3.i, SINGLETON(f3.i) AS col2
                 FROM f3
                 GROUP BY f3.i) AS dt1
                ON f2.i = SINGLECHK(dt1.col2)
         GROUP BY f2.i, dt1.i) AS dt2
        ON dt2.i = c1.i AND dt2.col2 = c1.i
WHERE f1.i = c1.i AND f1.i = SINGLECHK(dt2.col3)

In pslsgram.yi allow SINGLETON to return nullable regardless
of argument. It's not expected that user will use SINGLETON
explicitly in SQL but it does allow for easier testing of
flattening behavior.

In pslssflat.c
Track containing subsel/root as well as subsel to allow for seeing
when a join will span beyond the containing select.
In the PST_BOP handling we now use the tvrm maps of the subselect
and parent subselect to categorise the joins better and identify
whether the subselect's joinid should be used.
If the var nodes are contained in both from lists and both from
lists are referenced then we have a LEFT OUTER or INNER JOIN
depending on target list vs WHERE context respectivly. In these
case the join id will be used.
If all the vars are in the parent then these will float outward
and have no joinid. If all the vars are in the local set then these
will float inward and have no joinid.
The particular case where a deeper span is found does not trigger
an inner or outer to be set at this point as the actual join will
be recorded against the exported inner, maybe exported multiple
levels and the final export will be picked up later when the
updated BOP is seen in the tree higher up.
The routine psl_float_factors has the key parts to this change,
adding in the derived table range so that we can identify those
vars which were in the local from list but which at this point
have been changed to vars referencing the derived table - exported
so to speak. We need to make a distinction of these as depending
on their context they could be either a local or an external.
if the factor contains an outer then the dtvno var will be treated
as external. Otherwise, we switch it back to internal as it can
restrict the derived table closer to source.
Making the decision to float out means that the joinid should be
set whereas floating inward requires that the joinid be cleared
as the comparands can be sorted out within the derived table.

------------------------------------------------------------------------
r3420 | drivi01 | 2010-08-09 19:00:17 -0400 (Mon, 09 Aug 2010) | 5 lines

PORT change #506381 to 64-bit installer:

cazipxp.exe is no longer needed in Windows distribution manifest.

------------------------------------------------------------------------