------------------------------------------------------------------------ r3420 | drivi01 | 2010-08-09 19:00:17 -0400 (Mon, 09 Aug 2010) | 5 lines
PORT change #506381 to 64-bit installer:

cazipxp.exe is no longer needed in Windows distribution manifest.

------------------------------------------------------------------------
r3421 | stephenb | 2010-08-09 19:46:52 -0400 (Mon, 09 Aug 2010) | 42 lines

Bug Number: 124172

rns.start
     124172
     (GENERIC)
      When using dynamic SQL against most drivers; if a variable is placed in the select
      list, this may cause the DBMS to return an illegal data-type in the GCA tuple
      descriptor, producing unpredictable results in the driver, possibly a driver crash.
rns.end

Description of changes:
desc.start
The GCA tuple descriptor (GCA_TD_DATA) is fabricated from the top RESDOM node in the
parse tree generated by PSF. During parsing, if the parser sees a variable in a
dynamic select statement that makes it impossible to yet determine the type of
the result, it will delay type resolution and DB_NODT will substituted as the type.
This delay in type resolution will continue until a "PREPARE" statement is completely
parsed, and the types will not be resolved until the statement is executed and the
front-end provides actual values for the query variables.

The GCA protocol allows the client to request a tuple descriptor in response to a "PREPARE".
If the client requests a tuple descriptor, and the RESDOM still contains DB_NODT because
a type has not yet been resolved. This type will be returned in the tuple descriptor.
GCA stipulates that DB_NODT is an illegal type for a tuple descriptor, so the DBMS needs
to return something other than this.

This change adds logic that checks the value in the RESDOM at the last point before
we are about to return the tuple descriptor for a "PREPARE". If the type in the RESDOM
is still DB_NODT, we will make one final attempt to resolve the type using the RHS of
the parse tree. If this fails, we will substitute a "catch-all" type of DB_LTXT_TYPE
in the GCA_TD_DATA to ensure that DB_NODT is not sent back to the client.

Since the tuple descriptor for a "PREPARE" represents a "hint" rather than the final type,
our "best guess", or a fall-back type, is a reasonable substitution (given that we cannot
send DB_NODT). When the query is eventually executed the tuple descriptor that
represents the query results will be adjusted to contain the correct type using the
provided value for the variable.

Related Service Desk Issues:  145724

Build Considerations:  NONE
desc.end
------------------------------------------------------------------------