------------------------------------------------------------------------ r3427 | thogo01 | 2010-08-10 19:00:16 -0400 (Tue, 10 Aug 2010) | 65 lines
When a cursor is first opened, the cursor is assumed to be positioned in
front of the result set but it is not known if the result set is empty.
JDBC method isBeforeFirst() is required to return false when the result
set is empty.

JDBC driver needs to be able to detect the condition where it is unknown
if the result set contains any rows.  There is a row count which is
initially negative and is set to the full size of the result set once it
has been determined.  The row count will also be negative if some rows
have been fetched, but the end of the result set has not been reached.

A max row indicator is being added which is initially 0 and is set to the
highest row fetched by the driver.  This will be non-zero if any rows have
been fetched.  If the result set is known to be empty, the max row will be
0 and the row count will also be set to 0.  The problem condition occurs
when the max row is 0 (unknown number of rows or empty result set) and the
row count is negative (size of result is unknown).

When the problem condition exists, the driver must determine if there are
any rows in the result set. The easiest thing to do is pre-fetch the first
block of rows.  If the result set is empty, row count will be set to 0.
If the result set is not empty, max row will be be set greater than 0.  In
both cases, the current implementation of isBeforeFirst() will then return
the correct value.  The cursor must be returned to the original position
at the start of the result set.  This solution only applies to scrollable
cursors.

The scrolling result set classes now override isBeforeFirst() to check
for the problem condition, pre-fetch the first data block and reposition
the cursor.  With the first block of rows cached, positioning before the
result set is simply a logical operation (the physical cursor is not
moved) with very little overhead.  If the next cursor operation targets
a row in the first block of rows, there will be virtually no impact on
performance.  If the first block of rows must be discarded, then some
performance impact will be seen.

There is a related problem with the isLast() method.  When the cursor is
positioned at the highest row fetched from the result set, it may not be
known if it is actually the last row.  The driver can detect this
condition using the new max row indicator.

If the current row is not known to be the last row but it is the highest
row fetched, then the driver needs to determine if there are any more
rows in the result set.  The easiest thing to do is fetch the next block
of rows, but start with the current row so that the logical cursor does
not need to be repositioned (only scrollable cursors provide fetching
the current row).  If there are no more rows, then only the current row
will be returned.  If there are more rows, then the next block of rows
will be ready for processing.  If the next cursor operation targets a
row in the cached rows, then there will be virtually no impact on
performance.  If the next block of rows is discarded, then some
performance impact will be seen.

The solution described above is not applicable when the fetch size is
limited to 1 rows (only the current row will be fetched and no
information about subsequent rows will be determined).  In this case
the driver will have to fetch the next row to determine if any more
rows follow, and then scroll backwards to position on the original row.
This double fetch request will impact performance.

The scrolling resutl set classes now override isLast() to check for the
problem condition, perform the fetch operation to determine if there are
any more rows, and reposition the cursor if needed.

------------------------------------------------------------------------
r3428 | maspa05 | 2010-08-11 04:00:30 -0400 (Wed, 11 Aug 2010) | 52 lines

Review Board IP  http://reviewboard.ingres.prv/r/5324

Cross integration of change 506677

Bugs 123189 and 123960 occur because we're unable to distinguish between a
readonly database and a database opened for read-only access. A readonly
database is one created using createdb -r and has a couple of special features.
It only has one location and it has a writeable copy of its cnf file in
II_DATABASE. However due to the way its created there is a cnf file in its
data location but this has the paths of the original source database not
the readonly db itself.

There are two issues. In dm0c_open we assume the database is a readonlydb
if we're passed the DM0C_READONLY flag. Also in dm2t, in read_tcb, we check if
the access mode is DCB_A_READ and if so assume it's a readonly database and
that all the tables are in a single location. The bugs occur because there
are now cases - notable incrementally rollforward datatabases - where the
access needs to be readonly but the database isn't a readonlydb. We will
assume it is and look for the cnf in II_DATABASE and all tables in a single
location. If either of these is not true we get errors.

Added the flag DCB_S_RODB for use in DMP_DCB.dcb_status to indicate a
readonlydb. Used this flag as the test in dm0c_open and read_tcb.

Created various flags so that we can pass the information about whether this
is a readonlydb. This is based originally on the access flags in iidatabase
table in iidbdb.

Need to be able to do set this in the RCP as well as the DBMS as the function
count_opens calls dm0c_open. In order to achieve this added a series of flags
so that LGadd can set this information, to be retrieved by LGshow.

In order to add a new flag DM0L_RODB needed a new flag variable as the
existing one was using all the available bits. Added a parameter to
dm0l_opendb to support this.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------