------------------------------------------------------------------------ r3428 | maspa05 | 2010-08-11 04:00:30 -0400 (Wed, 11 Aug 2010) | 52 lines
Review Board IP  http://reviewboard.ingres.prv/r/5324

Cross integration of change 506677

Bugs 123189 and 123960 occur because we're unable to distinguish between a
readonly database and a database opened for read-only access. A readonly
database is one created using createdb -r and has a couple of special features.
It only has one location and it has a writeable copy of its cnf file in
II_DATABASE. However due to the way its created there is a cnf file in its
data location but this has the paths of the original source database not
the readonly db itself.

There are two issues. In dm0c_open we assume the database is a readonlydb
if we're passed the DM0C_READONLY flag. Also in dm2t, in read_tcb, we check if
the access mode is DCB_A_READ and if so assume it's a readonly database and
that all the tables are in a single location. The bugs occur because there
are now cases - notable incrementally rollforward datatabases - where the
access needs to be readonly but the database isn't a readonlydb. We will
assume it is and look for the cnf in II_DATABASE and all tables in a single
location. If either of these is not true we get errors.

Added the flag DCB_S_RODB for use in DMP_DCB.dcb_status to indicate a
readonlydb. Used this flag as the test in dm0c_open and read_tcb.

Created various flags so that we can pass the information about whether this
is a readonlydb. This is based originally on the access flags in iidatabase
table in iidbdb.

Need to be able to do set this in the RCP as well as the DBMS as the function
count_opens calls dm0c_open. In order to achieve this added a series of flags
so that LGadd can set this information, to be retrieved by LGshow.

In order to add a new flag DM0L_RODB needed a new flag variable as the
existing one was using all the available bits. Added a parameter to
dm0l_opendb to support this.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3429 | maspa05 | 2010-08-11 10:00:11 -0400 (Wed, 11 Aug 2010) | 25 lines

Review Board IP  http://reviewboard.ingres.prv/r/5335

Cross integration of change 506693

Output of query text to errlog.log is being truncated at 16384. Remove this
limit.

Amended uleformat to remove the setting of the length to a max of 16384.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------