------------------------------------------------------------------------ r3429 | maspa05 | 2010-08-11 10:00:11 -0400 (Wed, 11 Aug 2010) | 25 lines
Review Board IP  http://reviewboard.ingres.prv/r/5335

Cross integration of change 506693

Output of query text to errlog.log is being truncated at 16384. Remove this
limit.

Amended uleformat to remove the setting of the length to a max of 16384.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3430 | schka01 | 2010-08-11 11:45:09 -0400 (Wed, 11 Aug 2010) | 9 lines

The cast_length variable was declared as i4, but passed
to pst_node to construct a length 2 constant.  Since
the DB_ANYTYPE union is (as usual) not actually used,
big-endian hardware picks up the wrong half of the
length for the resulting parse tree.

The simplest cure is to just change the variable to i2.

------------------------------------------------------------------------