------------------------------------------------------------------------ r3433 | drivi01 | 2010-08-11 14:15:13 -0400 (Wed, 11 Aug 2010) | 19 lines
There's a few build problems on x64 Windows.  Some are related to the changes
submitted in the past couple of months and some are related to cygwin 1.7.

The following changes, fix the x64 build:

- Fixed ODBC driver compile problem due to undefined pird.
  This was fixed in the original port but somehow made it back into the code.
- Added msi.lib for license setup for SetupDriver Visual project.
- Fixed VERS not to build ICE and to build 64-bit only.
- Fix Jamrules to deal with flex compilation issues due to
  incorrect file format by adding u2d switch in the action (cygwin 1.7 related)
- Fixed numerous environmental issues in bldenv.bat.
	- CPU setting using architecture
	- SHELL settings
	- CAZIPXP needs to be removed
	- set DIFF variable for new version of cygwin 1.7 diff.exe to fix
          "p rcompare"

------------------------------------------------------------------------
r3434 | kiria01 | 2010-08-11 16:15:11 -0400 (Wed, 11 Aug 2010) | 9 lines

DESCRIBE INPUT not setting in list param types other than first.

Whilst the const list was descended, the bop context was
not available to set the type to. The INLIST has been
catered for explicitly to insure that all inlist params
have the same type.

------------------------------------------------------------------------