------------------------------------------------------------------------ r3435 | schka01 | 2010-08-11 17:00:11 -0400 (Wed, 11 Aug 2010) | 26 lines
DMF is not sufficiently thread-safe to deal with inserts,
updates, or deletes to a table in one thread and fetches
from the same table in a child thread, especially for
BTREE.  The child thread table can easily lose track of
its position and return "no more rows" prematurely.

The immediate cause of the problem is that the child
thread RCB isn't found by dm1b_rcb_update when it's time
to tell other session RCB's that things have changed.
Allowing rcb-update to find the child RCB isn't the
answer, though, because DMF assumes that concurrency
is controlled by locking, and updates the position
info without any mutexing or thread safety.  The
child RCB would not be affected by locking, since it's
part of the same session, and thus the DMF
assumptions are not valid.

Ideally, this would be fixed in DMF, but for now we
will simply prohibit plans that access the result
table or any of its indexes under an EXCH.  This
prohibition can be relaxed in a couple situations,
such as when there's a sort on top.  (The sort must
necessarily run its input sub-plan to completion
before it can produce any rows for updating.)

------------------------------------------------------------------------