------------------------------------------------------------------------ r3437 | schka01 | 2010-08-11 20:15:12 -0400 (Wed, 11 Aug 2010) | 10 lines
Dataldr was failing to generate a trailing nl delimiter
for the COPY list if there was a sequence column that wasn't
the last column in the field list.  Apparently the old
COPY front-end code managed to allow the unexpected
newline when running in FROM PROGRAM mode, but the new
COPY code does exactly what the copy list tells it
to do.  Fixed dataldr to generate the nl delimiter
at the end of the copy list.

------------------------------------------------------------------------