------------------------------------------------------------------------ r3437 | schka01 | 2010-08-11 20:15:12 -0400 (Wed, 11 Aug 2010) | 10 lines
Dataldr was failing to generate a trailing nl delimiter
for the COPY list if there was a sequence column that wasn't
the last column in the field list.  Apparently the old
COPY front-end code managed to allow the unexpected
newline when running in FROM PROGRAM mode, but the new
COPY code does exactly what the copy list tells it
to do.  Fixed dataldr to generate the nl delimiter
at the end of the copy list.

------------------------------------------------------------------------
r3438 | miket | 2010-08-11 22:20:18 -0400 (Wed, 11 Aug 2010) | 10 lines

Trac Number: #584

Bug Number: 122403

Description of changes:
Log and journal changes for encryption
(1) SD 146192 - format encrypted entries as hex in auditdb
(2) SD 146244 - for logging use a tuple compare routine that is not
    confused by encryption

------------------------------------------------------------------------