------------------------------------------------------------------------ r3438 | miket | 2010-08-11 22:20:18 -0400 (Wed, 11 Aug 2010) | 10 lines
Trac Number: #584

Bug Number: 122403

Description of changes:
Log and journal changes for encryption
(1) SD 146192 - format encrypted entries as hex in auditdb
(2) SD 146244 - for logging use a tuple compare routine that is not
    confused by encryption

------------------------------------------------------------------------
r3439 | coomi01 | 2010-08-12 09:00:18 -0400 (Thu, 12 Aug 2010) | 14 lines

Review Board IP  http://reviewboard.ingres.prv/r/5332

Insert select order by, may augument the resdom with defaulted values.
In this case we need to update the resdom reference in the sort list
appropriately.

However, because resdom idents in the sort list are unordered we
need to re-walk the list for each update. Plugging the new resdom
number in directly may lead to duplicate references. To avoid this
we put the numbers in as negative values. By the time we get to the
end of the update processing, the risk of duplicates is eliminated
and we may revert those negative values back to positives.

------------------------------------------------------------------------