------------------------------------------------------------------------ r3443 | miket | 2010-08-12 16:18:14 -0400 (Thu, 12 Aug 2010) | 8 lines
Trac Number: #584

Bug Number: 122403

Description of changes:
Minor fix for hex auditdb display for encrypted tables - missed test to
adjust buffer for hex output display.

------------------------------------------------------------------------
r3444 | schka01 | 2010-08-12 17:15:11 -0400 (Thu, 12 Aug 2010) | 16 lines

The fundamental problem here was LOB temps being
destroyed prematurely.  DMF doesn't know anything about
LOB temp lifetime, so it guesses, and sometimes the guess
is wrong.

The real fix would be some way for PSF / OPF / QEF to
communicate more about LOB temp lifetimes to DMF;  but
in lieu of that (architectural) change, a short-term
fix was done.  The fix was to a) teach the type resolver
to not attempt constant folding that results in a LOB
result, and b) mark all LOB-valued operators with the
NOTVIRGIN flag so that CX's make no attempt to reuse
LOB values from one iteration to the next.

------------------------------------------------------------------------