------------------------------------------------------------------------ r3444 | schka01 | 2010-08-12 17:15:11 -0400 (Thu, 12 Aug 2010) | 16 lines
The fundamental problem here was LOB temps being
destroyed prematurely.  DMF doesn't know anything about
LOB temp lifetime, so it guesses, and sometimes the guess
is wrong.

The real fix would be some way for PSF / OPF / QEF to
communicate more about LOB temp lifetimes to DMF;  but
in lieu of that (architectural) change, a short-term
fix was done.  The fix was to a) teach the type resolver
to not attempt constant folding that results in a LOB
result, and b) mark all LOB-valued operators with the
NOTVIRGIN flag so that CX's make no attempt to reuse
LOB values from one iteration to the next.

------------------------------------------------------------------------
r3445 | loera01 | 2010-08-12 17:45:26 -0400 (Thu, 12 Aug 2010) | 15 lines

The ODBC driver maintains error messages in a structure named SQLCA_MSG_TYPE.
The static SQLERM field stored the error message, and had a maximum length of
256.  I changed SQLERM to a dynamic pointer and added SQLLEN to store the
message length.  The length had previouly been stored in the first byte of
SQLERM, allowing no more than 255 in length, and overflowing without error
for larger lengths.

The ErrInsertSqlcaMsg() function allocates SQLERM according to the string
length of the error message passed in.  The routines PopSqlcaMsg() and
FreeSqlcaMsgBuffers() free SQLERM.  The function GetDiagRec() had used a
static buffer for calling the routine ErrGetSqlcaMessage(), but since a
dynamic buffer is now required, the length needs to be known in advance.
I created ErrGetSqlcaMessageLen() to fetch the message length.

------------------------------------------------------------------------