------------------------------------------------------------------------ r3445 | loera01 | 2010-08-12 17:45:26 -0400 (Thu, 12 Aug 2010) | 15 lines
The ODBC driver maintains error messages in a structure named SQLCA_MSG_TYPE.
The static SQLERM field stored the error message, and had a maximum length of
256.  I changed SQLERM to a dynamic pointer and added SQLLEN to store the
message length.  The length had previouly been stored in the first byte of
SQLERM, allowing no more than 255 in length, and overflowing without error
for larger lengths.

The ErrInsertSqlcaMsg() function allocates SQLERM according to the string
length of the error message passed in.  The routines PopSqlcaMsg() and
FreeSqlcaMsgBuffers() free SQLERM.  The function GetDiagRec() had used a
static buffer for calling the routine ErrGetSqlcaMessage(), but since a
dynamic buffer is now required, the length needs to be known in advance.
I created ErrGetSqlcaMessageLen() to fetch the message length.

------------------------------------------------------------------------