------------------------------------------------------------------------ r3447 | schka01 | 2010-08-12 22:45:13 -0400 (Thu, 12 Aug 2010) | 20 lines
Coupon replace wasn't being sufficiently careful about
ensuring that the short-term part of the new coupon in
the new row is zero.  In particular, for a "no change"
situation, the new coupon might have stuff in the
short term part from a previous get, and that stuff has
to be cleared before the row is written.

MVCC has nothing to do with the problem per se, it's
just the easiest way to provoke it since MVCC will
attempt to undo changes back to a consistent-read
state.  When the undo sees what looks like a non-
logged change (to the short term part of the coupon),
it goes nuts and rejects the undo with a string of
errors.

In addition, improve the error testing done by
trace point DM722, which verifies that coupons in a
row just inserted or updated are correct.

------------------------------------------------------------------------