------------------------------------------------------------------------ r3450 | horda03 | 2010-08-13 11:00:15 -0400 (Fri, 13 Aug 2010) | 39 lines
Review Board IP  http://reviewboard.ingres.prv/r/5371

Cross integration of change 506743

IMA table ima_server_sessions truncates return value from column client_info which is defined as:

varchar(64) not null not default is 'exp.scf.scs.scb_client_info'

This IMA parameter refers to a DBMS session's scb->scb_sscb.sscb_ics.ics_gw_parms.
This is defined when the session is initiated and is set to the GCA_GW_PARM value
sent from the FE. This variable text message is limited by the FE to IIsqpMAX_PARM_SIZE
(1000) characters, and contains the various types of parameters
that are sent to a gateway (or in this case the DBMS) as a string of commands

e.g.

user='ingres',host='horda0393',tty='3',pid=25881,conn='a1234567890123456789012345678'

If the length of the string is longer than the definition of client_info, then only the
first "client_info" length characters are displayed.

I believe 1000 is being overzealous, so upping hte limit to 200.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------