------------------------------------------------------------------------ r3460 | hanje04 | 2010-08-13 20:15:10 -0400 (Fri, 13 Aug 2010) | 5 lines
Follow change to 505551, missed the condition where only renamed instances
are installed. It's rare but it could happen.

------------------------------------------------------------------------
r3461 | loera01 | 2010-08-13 22:15:11 -0400 (Fri, 13 Aug 2010) | 26 lines

The Unicode version of the error functions, SQLGetDiagRecW(),
SQLGetDiagFieldW() and SQLGetErrorW() relied on a 512-character buffer to
converting Unicode strings back to multi-byte.  I revised the code so that a
dynamic buffer was allocated if the message length was greater
than SQL_MAX_MESSAGE_LENGTH - 50 (462).  The50 bytes account for the error
message header that is inserted, which currently takes about 30 bytes.

The internal GetDiagRec() routine had unilaterally used a dynamic work buffer
for error messages.  I revised GetDiagRec() along the same lines as the
Unicode routines.  In practice, the chances of an error message being longer
than 512 characters is rare, so there is no need to continually allocate and
free message buffers.

I sharpened up the code a bit in other areas.  The psqlca pointer needed to be
initialized to NULL because conditional code depended on a NULL setting.  The
GetDiagRec() used FreeDbc() to free the statement, connection and descriptor
handles, which currently works because the same critical seciton object is
used for all three.  Nevertheless, it's clearer to mark the difference, as
this could change in the future.

I also changed the SQLLEN field in the SQLCA_MSG_TYPE block to SQL_MSG_LEN,
since SQLLEN is a definition macro and caused the Devstudio debugger to list
the field as "SQLINTEGER".  Other compilers may not be as forgiving of the
SQLLEN definitinon as the MS C compiler.

------------------------------------------------------------------------